requirements.txt
# Requirements for image to pdf
PyPDF2==1.26.0
Pillow==8.4.0
img2pdf==0.4.1

